tweaking
and getting stuck, in my example yaml I already used a wildcard
not just as a suffix, but in between. And that would disable me checking
just on the basis of the expressions whether sets are disjoint, or
at the very least make that a whole lot harder, if not impossible.

But if I disable the infix wildcard, it does mean my YAML, even my example
YAML, get's a lot longer.